moved cordinates helpers to utils file

+++ apps/backend/src/games/games.gateway.ts
@@ -11,6 +11,7 @@ import {
 import { GuessesService } from 'src/guesses/guesses.service';
 import { MembersService } from 'src/members/members.service';
 import { RoundsService } from 'src/rounds/rounds.service';
+import { getDistance } from 'src/util';
 import { GamesService } from './games.service';
 
 @WebSocketGateway({
@@ -50,6 +51,9 @@ export class GamesGateway implements OnGatewayConnection, OnGatewayDisconnect {
 
   async handleDisconnect(client) {
     const memberId = this.clientToMember.get(client.id);
+    if (!memberId) {
+      return;
+    }
     const member = await this.membersService.removeAt(memberId);
     await this.updateGameState(member.gameId);
 
@@ -159,20 +163,4 @@ function parseCordinates(val: string) {
   const rawLatLng = val.split(',');
   const latLng = { lat: Number(rawLatLng[0]), lng: Number(rawLatLng[1]) };
   return latLng;
-}
-
-function getDistance(lat1, lon1, lat2, lon2) {
-  const R = 6371; // Radius of the earth in km
-  const dLat = deg2rad(lat2 - lat1);
-  const dLon = deg2rad(lon2 - lon1);
-  const a =
-    Math.sin(dLat / 2) * Math.sin(dLat / 2) +
-    Math.cos(deg2rad(lat1)) * Math.cos(deg2rad(lat2)) * Math.sin(dLon / 2) * Math.sin(dLon / 2);
-  const c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-  const d = R * c; // Distance in km
-  return d;
-}
-
-function deg2rad(deg) {
-  return deg * (Math.PI / 180);
 }
